WIKI Apothecary (Background)
Added the Apothecary background as a character customization option. When this file is included, the option can be located under "Build->Background" and selected from the drop-down list. This background grants most features it describes to its player and allows the player to select the features it requests they decide on. The pane describes the background according to what's presented on the wiki. Background title appears on character sheet preview along with other information.

As of this version, the background does not automatically grant equipment or money and does not write its own description or main feature on the character sheet preview; and although the background grants proficiency with the herbalism kit, the kit is not available from the items list. All items granted but not on the list must be written manually to appear on the sheet.